casting int to float so there's no loss of data when calculating cpu usage

+++ src/processor.cpp
@@ -6,6 +6,6 @@ float Processor::Utilization()
   float utilization{0};
   auto totalJiffies = LinuxParser::Jiffies();
   auto activeJiffies = LinuxParser::ActiveJiffies();
-  utilization = activeJiffies / totalJiffies;
+  utilization = static_cast<float>(activeJiffies) / totalJiffies;
   return utilization;
 }
